requirements.txt
# Essential Packages
fastapi==0.111.0
pydantic==2.7.4
uvicorn==0.30.1

# Commonly Used Packages
httpx==0.27.0
python-multipart==0.0.9
requests==2.32.3


# Optional Packages (Include if Needed)
anyio==4.4.0
black==24.8.0
flake8==7.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
isort==5.13.2
pydantic-extra-types==2.8.1
pydantic-settings==2.3.3
typing_extensions==4.12.2
rich==13.7.1
python-dotenv==1.0.1
websockets==12.0

# Development and Additional Packages
click==8.1.7
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi-cli==0.0.4
orjson==3.10.4
pydantic-core==2.18.4
ujson==5.10.0
uvloop==0.19.0
watchfiles==0.22.0
typer==0.12.3
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.23.8
pytest-mock
